pyproject.toml
[tool.poetry]
name = "infinitecopy"
version = "0.1.0"
description = "Simple clipboard manager"
license = "GPL-3.0-or-later"
authors = ["Lukas Holecek <hluk@email.cz>"]
include = ["infinitecopy.png", "qml/*"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
PySide6-Essentials = "^6.5.0"
PyGObject = [
  {version = "^3.42.0", markers = "platform_system == 'Linux'"}
]

[tool.poetry.dev-dependencies]
Nuitka = "^1.5.4"
zstandard = "^0.20.0"
imageio = [
  {version = "^2.22.4", markers = "platform_system != 'Linux'"}
]
pillow = "^9.3.0"
pyright = "^1.1.296"
pylint = "^2.17.1"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
infinitecopy = "infinitecopy.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
ignore-paths = [
    "^tests/",
]
jobs=0

[tool.pylint."messages control"]
disable = [
    "broad-exception-caught",
    "invalid-name",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "no-name-in-module",
    "redefined-outer-name",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "too-many-return-statements",
    "wrong-import-position",
]
